Refactored App.js into seperate components.
Created Components GoalInput and GoalItem.
