+++ react-native-the-practical-guide/section-2-react-native-basics/CourseGoals/App.js
@@ -1,23 +1,17 @@
 import { useState } from 'react';
 import { 
   StyleSheet, 
-  Text, 
   View, 
-  Button, 
-  TextInput, 
-  ScrollView,
   FlatList 
 } from 'react-native';
 
+import GoalItem from './components/GoalItem';
+import GoalInput from './components/GoalInput';
+
 export default function App() {
-  const [enteredGoalText, setEnteredGoalText] = useState('');
   const [courseGoals, setCourseGoals] = useState([]);
 
-  function goalInputHandler(enteredText) {
-    setEnteredGoalText(enteredText);
-  }
-
-  function addGoalHandler() {
+  function addGoalHandler(enteredGoalText) {
     setCourseGoals((currentCourseGoals) => [
       ...currentCourseGoals,
       { text: enteredGoalText, id: Math.random().toString() },
@@ -26,23 +20,14 @@ export default function App() {
 
   return (
     <View style={styles.appContainer}>
-      <View style={styles.inputContainer}>
-        <TextInput style={styles.textInput} 
-                   placeholder='Your course goal!' 
-                   onChangeText={goalInputHandler} />
-        <Button title='Add Goal' onPress={addGoalHandler}/>
-      </View>
+      <GoalInput 
+        onAddGoal = {addGoalHandler}
+      />
       <View style={styles.goalsContainer}>
         <FlatList 
           data={courseGoals} 
           renderItem={(itemData) => {
-            return (
-              <View style={styles.goalItem}>
-                <Text style={styles.goalText}>
-                  {itemData.item.text}
-              </Text>
-            </View>
-            )
+            return <GoalItem text={itemData.item.text}/>
           }}
           keyExtractor={(item, index) => {
             return item.id;
@@ -59,32 +44,7 @@ const styles = StyleSheet.create({
     paddingTop: 50,
     paddingHorizontal: 16
   },
-  inputContainer: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 24,
-    borderBottomWidth: 1,
-    borderBottomColor: '#cccccc'
-  },
-  textInput: {
-    borderWidth: 1,
-    borderColor: '#cccccc',
-    width: '70%',
-    marginRight: 8,
-    padding: 8
-  },
   goalsContainer: {
     flex: 5,
-  },
-  goalItem: {
-    margin: 8,
-    padding: 8,
-    borderRadius: 6,
-    backgroundColor: '#5e0acc',
-  },
-  goalText: {
-    color: 'white'
   }
 });
